filter: compare as integers to make sure we're within the right bounds.
Fixes CID 1189413: Unsigned compared against 0 (NO_EFFECT)

+++ gr-filter/lib/pfb_synthesizer_ccf_impl.cc
@@ -220,15 +220,15 @@ namespace gr {
       gr::thread::scoped_lock guard(d_mutex);
 
       if(map.size() > 0) {
-        unsigned int max = (unsigned int)*std::max_element(map.begin(), map.end());
-        unsigned int min = (unsigned int)*std::min_element(map.begin(), map.end());
-        if((max >= d_twox*d_numchans) || (min < 0)) {
+        int max = *std::max_element(map.begin(), map.end());
+        int min = *std::min_element(map.begin(), map.end());
+        if((max >= static_cast<int>(d_twox*d_numchans)) || (min < 0)) {
           throw std::invalid_argument("pfb_synthesizer_ccf_impl::set_channel_map: map range out of bounds.\n");
         }
         d_channel_map = map;
 
         // Zero out fft buffer so that unused channels are always 0
-        memset(d_fft->get_inbuf(), 0,d_twox*d_numchans*sizeof(gr_complex));
+        memset(d_fft->get_inbuf(), 0, d_twox*d_numchans*sizeof(gr_complex));
       }
     }
